Don't store invalid exit status in case of waitpid() error
io_context_ref should not store the -1 exit status in case waitpid()
failed, because it could overwrite the proper exit status from a previous
waitpid() call.

(especially because the -1 is not even coming from waitpid(), but purely
an internal fallback value required because of the completion handler's
function signature)

This can happen for example when a program is checking is_running(),
which calls waitpid() internally, reaps the child process and stores the
exit status, before the handle_sigchld callback is invoked. It can be
delayed quite a bit since it is queued in the io_service which may have
other work to process first.

This feels like an easy fix, but perhaps not the best solution.
A better solution might be to rework boost::process to stop calling
waitpid() for a certain child PID after it reaped it once.
Then it would also be better protected against PID re-use issues.